refactor: extract class from calculator
